Fixes comment: https://github.com/apache/incubator-kie-tools/pull/2167#discussion_r1516835107
